connector/test_talk/consumer/requirements.txt
kafka-python
google
protobuf